Date: 13, november 2025 Lesson number: 1. Goal for today (C i)  State what your plan said you should complete during this session.  My goal was to produce a functioning piece of code that allowed me to input gender, date of birth and able to check the data. I also wanted to add more inputs for the form to get closer to fully having all necessary inputs.
2. Work completed (C iii)
 Write one or two sentences summarising what you achieved in this session.
I managed to finally allow data to be stored and viewed using json which was what was this week was what we we meant to what we were meant to do this weak to accomplish what was we were meant to do this weak.

3. Did you follow your plan? (C iii and C iv)
Yes and no:
Yes completely finishing what we were meant to finish for the week that the finish of the json file was due which was today I think.

4. Next step (C i)
 List one or two tasks you intend to complete during the next lesson.
 I need to add a few more boxes for inputs and then change the layout to be more pleasing.